requirements.txt
addict==2.4.0
aiohttp==4.0.0a1
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
autoprop==4.0.2
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
bleach==4.1.0
cairocffi==1.3.0
CairoSVG==2.5.2
centerline==0.6.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorlog==6.6.0
commonmark==0.9.1
cssselect2==0.5.0
cycler==0.11.0
Cython==3.0.0a10
-e git+git@github.com:jjmontesl/ddd.git@b3f38e281bb80750a37b628646f7b8cbb878db0a#egg=ddd
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
executing==0.8.3
freetype-py==2.2.0
GDAL==3.0.4
geographiclib==1.52
geojson==2.5.0
glooey==0.3.6
idna==3.3
imageio==2.16.1
importlib-metadata==3.10.1
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==8.0.0rc0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.11.1
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.11.2
jupyterlab-widgets==3.0.0rc0
kiwisolver==1.4.0
lark-parser==0.12.0
lxml==4.8.0
mapbox-earcut==1.0.0
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
meshio==5.3.4
mistune==0.8.4
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.3
multidict==4.7.6
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
noise==1.2.2
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
open3d==0.13.0
osmium==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
Pint==0.18
portion==2.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycatenary==0.4.0
pycollada==0.7.2
pycparser==2.21
pycsg==0.3.3
pyembree==0.2.11
pyGeoTile==1.0.6
pyglet==1.5.23
Pygments==2.11.2
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.7
pypng==0.0.21
pyproj==2.6.1.post1
pyrender==0.1.45
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-fcl==0.6.1
python-socketio==5.5.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.0.0b1
requests==2.27.1
rich==12.0.0
Rtree==0.9.7
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.5.4
Send2Trash==1.8.1b0
Shapely==1.8.1.post1
signature-dispatch==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.2.0
svg.path==5.0.1
svgpath2mpl==1.0.0
svgpathtools==1.4.4
svgwrite==1.4.1
sympy==1.10.1
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.3.16
tinycss2==1.1.1
tomlkit==0.10.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
triangle==20220202
trimesh==3.10.5
typeguard==2.13.3
typing-extensions==4.1.1
urllib3==1.26.9
vecrec==0.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
widgetsnbextension==4.0.0rc0
xatlas==0.0.6
xxhash==3.0.0
yarl==1.7.2
zipp==3.7.0
